fix: start osd if org.freedesktop.Notifications is not owned
When multiple dbus services that announced owning the same name,
none of them will be started when the name is needed.
It results that dde-osd will not run if other package that will own
org.freedesktop.Notifications like KDE Plasma desktop is installed.

dbus-broker warns:
  Ignoring duplicate name 'org.freedesktop.Notifications' in service file '/usr/share/dbus-1/services/com.deepin.dde.freedesktop.Notification.service'

This patch makes dde-osd always be started when org.freedesktop.Notifications
is not owned.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1711001